[DEV-154] rename: goal 컴포넌트 사용되는 곳으로 이동
